% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macromodels.R
\docType{package}
\name{macromodels}
\alias{macromodels}
\alias{macromodels-package}
\title{macromodels: Introductory graduate-level macroeconomics.}
\description{
This is a companion package for learning introductory graduate-level
macroeconomics. It includes ...
}
\references{
Data sources
\itemize{
\item Feenstra, Robert C., Robert Inklaar and Marcel P.
Timmer (2015), "The next generation of the Penn World Table," \emph{American
Economic Review}, 105(10), 3150-3182, available for download at
\href{http://www.ggdc.net/pwt}{www.ggdc.net/pwt}
}

Textbooks
\itemize{
\item Jones, Charles I. and Dietrich Vollrath (2013) \emph{Introduction to
Economic Growth} 3e. W. W. Norton.
\item Romer, David (2011) \emph{Advanced
Macroeconomics} 4e. McGraw-Hill.
}

Journal articles
\itemize{
\item Mankiw, N. Gregory, David Romer and David N. Weil (1992)
"A contribution to the empirics of economic growth," \emph{Quarterly Journal of
Economics}, 107(2), 407-437. \url{http://www.jstor.org/stable/2118477}
}
}
