% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Simulate State Space Model}
\usage{
run_simulation(h, g = NULL, x0, t)
}
\arguments{
\item{h}{function for state equation}

\item{g}{function for output equation}

\item{x0}{initial state vector}

\item{t}{integer, length of simulation}
}
\value{
output matrix
}
\description{
x_{t+1} = h(x_t)
y_{t} = g(x_t)
}
